refactor(search): final renaming of components
Final renaming of components
