+++ src/components/UI/icons/IconSearch.js
