+++ src/components/UI/icons/RemoveSearch.js
